test: 소수를 추출할 수 있음에 따른 Parser 테스트 수정

+++ src/test/java/calculator/domain/ParserTest.java
@@ -3,6 +3,7 @@ package calculator.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import java.math.BigDecimal;
 import java.util.List;
 import org.junit.jupiter.api.Test;
 
@@ -10,25 +11,25 @@ public class ParserTest {
     @Test
     void 문자와_숫자_섞인_문자열에서_숫자만_추출되어_정수로_파싱된다() {
         Parser parser = new Parser(new String[]{"nn1", "a2$", "3$$"});
-        List<Integer> numbers = parser.parseIntegers();
+        List<BigDecimal> numbers = parser.parseIntegers();
 
-        assertThat(numbers).contains(1, 2, 3);
+        assertThat(numbers).contains(BigDecimal.valueOf(1), BigDecimal.valueOf(2), BigDecimal.valueOf(3));
     }
 
     @Test
     void 숫자가_하나도_없으면_0으로_치환된다() {
         Parser parser = new Parser(new String[]{"", "abc", "$$"});
-        List<Integer> numbers = parser.parseIntegers();
+        List<BigDecimal> numbers = parser.parseIntegers();
 
-        assertThat(numbers).contains(0, 0, 0);
+        assertThat(numbers).contains(BigDecimal.valueOf(0), BigDecimal.valueOf(0), BigDecimal.valueOf(0));
     }
 
     @Test
     void 문자열_내_여러_숫자는_연결되어_하나의_정수로_해석된다() {
         Parser parser = new Parser(new String[]{"qwer1er3", "dk0dk1"});
-        List<Integer> numbers = parser.parseIntegers();
+        List<BigDecimal> numbers = parser.parseIntegers();
 
-        assertThat(numbers).contains(13, 1);
+        assertThat(numbers).contains(BigDecimal.valueOf(13), BigDecimal.valueOf(1));
     }
 
     @Test
